# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dopamine Market API
  description: >
    Синтетический маркетплейс/сервис доставки — портфолио-проект.
    Монетизация: фиксированная плата 10₽ за оформленный заказ (INV-01).
    Полностью синтетический каталог и курьеры, симуляция доставки —
    таймер-стейт-машина (см. docs/prd/PRD.md, ADR-005, ADR-009).
  version: "1.0.0"
  contact:
    name: Dopamine Market (портфолио-проект)
servers:
  - url: http://localhost:8080
    description: Локальная разработка
  - url: https://api.dopamine-market.example.com
    description: Продакшн (заглушка домена)
...
paths:
...
  /payments/{order_id}:
    get:
      tags: [payment]
      summary: Получение статуса платежа по ID заказа
      operationId: getPaymentByOrderId
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          $ref: "#/components/responses/NotFound"
  /payments/webhook/mock:
    post:
      tags: [payment]
      summary: Вебхук тестового Mock-провайдера
      operationId: handleMockWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_payment_id, order_id, status]
              properties:
                provider_payment_id:
                  type: string
                order_id:
                  type: string
                  format: uuid
                status:
                  type: string
                  enum: [succeeded, failed]
                failure_reason:
                  type: string
      responses:
        "200":
          description: Вебхук успешно обработан
        "400":
          $ref: "#/components/responses/BadRequest"
  /payments/webhook/yookassa:
    post:
      tags: [payment]
      summary: Вебхук YooKassa API (ADR-006)
      operationId: handleYooKassaWebhook
      security: []
      responses:
        "200":
          description: Вебхук успешно обработан
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    Payment:
      type: object
      required: [id, order_id, provider, provider_payment_id, amount_rub, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        provider:
          type: string
        provider_payment_id:
          type: string
        amount_rub:
          type: string
          example: "10.00"
        status:
          type: string
          enum: [pending, succeeded, failed]
          example: "pending"
        confirmation_url:
          type: string
        failure_reason:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...